Make pointer types in messages be 64bit in memory
Right now we always assume pointers are stored as SPICE_ADDRESS,
i.e. 64bit, independent on the size sent on the network.
This is required for 64bit architectures of course, but slightly overkill
on 32bit architectures, so needs fixing when all SPICE_ADDRESS elements
can be made internal.

+++ common/messages.h
@@ -442,8 +442,8 @@ typedef struct SpiceMsgcTunnelAddGenericService {
     uint32_t id;
     uint32_t group;
     uint32_t port;
-    uint32_t name;
-    uint32_t description;
+    uint64_t name;
+    uint64_t description;
 } SpiceMsgcTunnelAddGenericService;
 
 typedef struct SpiceMsgcTunnelAddPrintService {
